More mountpoint fixes
Copied from Perforce
 Change: 13140

+++ rpmconf/Install/Util/utilfunc.sh
@@ -526,7 +526,9 @@ removeExistingInstall() {
 		umount /opt/zimbra/amavisd/tmp > /dev/null 2>&1
 		MOUNTPOINTS=`mount | awk '{print $3}' | grep /opt/zimbra`
 		for mp in $MOUNTPOINTS; do
-			/bin/rm -rf ${mp}/*
+			if [ x$mp != "x/opt/zimbra"]; then
+				/bin/rm -rf ${mp}/*
+			fi
 		done
 		find /opt/zimbra -type f -exec /bin/rm -f {} \;
